Clean up build errors

+++ source/drivers/memory.h
@@ -2,4 +2,4 @@
 #include "source/updater/updater.h"
 
 err_t writeBytesToMemory(ImageBaseAddress image, uint32_t start_address,
-                         uint32_t* data, uint32_t data_size);
+                         uint8_t* data, uint32_t data_size);
